gateway_service/requirements.txt
flask
Flask-Smorest
Flask-JWT-Extended
marshmallow
python-dotenv
gunicorn
requests
redis
pika